requirement.txt
ctranslate2==3.17.1
transformers==4.34.0
numpy==1.25.0
flash-attn==2.2.2
dm-tree==0.1.8
accelerate==0.22.0
deepspeed==0.10.3
peft==0.3.0
torchtyping==0.1.4
jsonlines
datasets==2.14.5
sentencepiece==0.1.99
huggingface-hub==0.17.3
